+++ NEWS.md
@@ -1,3 +1,6 @@
+# beisics ......
+* Excel styling now available when saving a workbook using the openxlsx package. Use add_beis_header_style to style the column headers and add_beis_body_style to style the body.
+
 # beisics 0.0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.
@@ -5,3 +8,4 @@
 * beis_colours() update to the latest colour scheme and also given an alternative palette. Alternative palette is selectable via the defaul = FALSE argument.
 
 * theme_beis now includes a call to beis_colours for scale_fill_manual and scale_fill_colour and included the option to use the alternative palettte or the distinct colour palette.
+